Add rake task to anonymise emails (#165)
https://eaflood.atlassian.net/browse/WC-525

We often need to test stuff against production data. If when testing we complete a renewal then the service will send an email to the actual user. We could break the link to our email server within the environment under test, however this could lead to unrelated errors occurring during testing because an email failed to send.

So we really need someway that having restored production data ready for debugging/testing, we can then anonymise the email addresses so they will never actually get sent to a real inbox.

This change implements an anonymise email rake task. It will iterate through all users, then update the contact and account email in all registrations and transient_registrations linked to the user, before updating the user account itself.

The change is relatively large, because it also includes changes to make rake tasks more scalable in the future.

It encompasses starting to move some of the similar rake task functionality into the `lib`, from where it can be better managed. We followed advice in [How to structure a Ruby library](https://ryanlue.com/posts/2017-03-03-how-to-structure-a-library) for where to put things, hence we now have

- lib
  - db
    - db.rb
    - anonymise_email.rb
  - db.rb

Doing this also solves 2 issues

- Rubocop flagging overly large blocks in the rake file
- Methods defined in rake tasks are simply added to the global namespace, hence our method names could be clashing (see https://kevinjalbert.com/defined_methods-in-rake-tasks-you-re-gonna-have-a-bad-time/)