Add timings and alter output in anonymise email task (#170)
The rake task will take sometime to run, hence it is likely we would leave it running unattended.

For future planning, and to help determine if further optimisation is needed it would be useful to know approximately how long it takes to run.

Hence this tweak to include timings in the output of the rake task.

We also tweaked the output to improve performance. Displaying the details of the changes for each account can be useful for debug purposes, but will slow things down (if only marginally).

So we tweaked the task so now if we want full output, we state we want to debug, else by default the anonymiser will print a period for each page (100) of users it processes, so we have some indication that things are progressing.