Fix blocked actions post grace window (#169)
https://eaflood.atlassian.net/browse/WC-512

During testing of the grace window it was spotted that if a renewal was expired, and it matched during conviction checks, then it could not be continued because the button to access convictions approvals did not display.

On double checking we found built into the view the check `ACTIVE?`, i.e. if the renewal was not ACTIVE then it could not be approved.

We needed to alter this check to call `may_renew?` however this would also mean ensuring we had a link to the corresponding registration.

We found when looking into an issue in the waste_carriers_engine that we had an assumption we could access the registration from the metaData object, irrespective of whether the metadata was on a registration or a transient_registration.

The actual case is that you can only do it if the metaData is on a registration, hence `may_renew?` will fail if called on a metaData linked to a transient registration.