Add ability to skip already anonymised emails (#173)
The anonymiser currently takes so long to complete that it's progress is often cut short. However it seems a waste of effort to start from the beginning anonymising emails that have already been anonymised so this change adds a test of the current user email address.

If it ends with '@example.com' the anonymiser skips updating the user and associated records and moves onto the next one.