Transfer a registration to a new account (#161)
* Transfer a registration to a new account

https://eaflood.atlassian.net/browse/WC-61

Now that we can transfer a registration to an existing account, we should also be able to transfer it to a new one.

When the registration transfer form is submitted, if there is no account with the email address used, immediately create a new account and transfer the registration.

The new account should then receive an email notifying them of the change and inviting them to set their password.

* Create a new user if no matching one exists

Use the devise_invitable gem to generate a new ExternalUser when we can't find an existing one. This gem includes views and emails, but we skip using them for now as we'll need to build our own.

The status returned is also inaccurate - this will be fixed shortly.

* Refactor tests so service isn't called twice

Previously some specs called `transfer_to_email` before running the spec, and then called it a second time to test what it returns. This doesn't accurately represent how the service would be used as this should only called once, and calling it twice could produce a different result (for example, if the user doesn't exist at first but is created the second time). So instead we call it in a before block but save the result, and then test that result instead of rerunning it.

* Add more tests for new users

Make sure that emails are being sent and the account_emails are being updated.

* Refactor spec to call transfer_to_user every test

Initially we tried to reduce duplication by running transfer_to_user as little as possible and then testing that result multiple times. However this results in the tests being less accurate, and makes it harder to test changes before and after the method is called. So instead we call it in each test at the point at which it's required.

* Return :success_new_user when new user is created

If the RegistrationTransferService has to create a new user, the returned value should indicate that a new user was created. This means we will be able to display a different result depending on whether the user already or exists or is new.

* Form submits successfully when new user created

Update the RegistrationTransferForm to consider :success_new_user a valid response from the service.

* Set up email for transfers to new accounts

Configure the mailer to add a new confirmation email, this time to be sent to accounts created by the transfer.

Currently the content is the same as the existing account transfer email.

* Update service to send different email for new users

When the service creates a new user, send them the email specifically for new users.

* Include URL to accept invite in email

When we invite a new user, devise_invitable generates a new token which is used to accept the invitation. We pass this token to the mailer to generate the correct URL to accept the invitation in the front-office.

* Update config to match front-office

This won't have any effect in the back-office, but we should keep config consistent across the apps. We use this option to automatically sign the new user in after they set their password, rather than having them enter it for a third time.

* Disable invitation routes

We aren't using any of the pages devise_invitable provides to invite new users or accept invitations, as the invitation process is just part of the registration transfer process. So we disable the routes to avoid accidental usage.

* Add helper to check if an account has just been invited

This method checks to see if the account has been invited and that invitation has not yet accepted.

We can use this to display a different message to the NCCC user depending on the result.

* Display different success message if account is new

When the NCCC user transfers a registration, they will see a success page. This page now displays different text depending on whether or not the account is new, so they can tell the user what the next steps are.

* Button on success page redirects to backend

While adding the link to this feature in the backend (https://github.com/DEFRA/waste-carriers-frontend/pull/176) I realised there was no clear path back to where the user came from after transferring the registration. So instead the button now links back to the backend dashboard, filtered to show the now-updated registration.