Add new page for user management (#180)
* Add new page for user management

https://eaflood.atlassian.net/browse/WC-495

Agency super users and finance super users should be able to manage back office users - for example, adding or removing users, or changing what permissions they have.

As a starting point for this, we need to have a page which provides an overview of user management which super users can access. This page should be in the menu.

* Create user index page

Set up a new UsersController with an index action, accessible from "/bo/users".

* Block people who aren't logged in from accessing the page

If a user is not logged in, they will be redirected to the sign-in page.

* Add new ability :manage_back_office_users

Agency super admins and finance super admins are concerned with different types of users, but they both need to access this screen. So we introduce a new ability to check if they are involved in any sort of user management.

* Only users who can :manage_back_office_users can view page

Everyone else (non-super users) should be redirected to the permissions error page.

* List back office users on page

List all the back office user emails, plus their role (in a humanised format).

* Add 'manage users' link to top level nav

This is only visible to superusers who are able to manage users, so regular users won't see it.

* Paginate list of users

We have a lot of back office users, so pagination makes this list much more manageable. We use Kaminari to paginate the list. It shares the same config as the dashboard pagination, showing 20 results per page.

Also change the table to be a full-width column so we have space for the pagination controls.