Add rake task to sync internal users (#182)
https://eaflood.atlassian.net/browse/WC-524

When the service went live on September 19 we ran a JS script to create users in the back office based on users that already existed in the old back end.

As we do not currently have the ability to maintain users in the back office (see WC-495) this still remains the only means to create users.

Having spoken with the NCCC team lead, just being able to re-sync the back office users with those in the backend and create any new ones would suffice.

Hence this change which adds a rake task to do just that. It

- compares both the backend **admins** and **agency_users** collections with the **back_office_users** collection
- any it does not find in **back_office_users** it creates, with the exception of agency user's that already have an admin user that has been synced (to avoid duplicates)
- any matches it finds it compares the role assigned, and updates the back office user if they do not match

The users it creates will have the same password so there is no need to set a new one by users.

For reference there is also a future intention to enable access to admin users to run this function, hence some of the design decisions. However this will come in future PR's.