Allow super users to migrate users from backend (#187)
* Allow super users to migrate users from backend

https://eaflood.atlassian.net/browse/WC-495

Super users can create users in the backend. However, as these are currently stored in different DB collections, a new backend user won't automatically have access to the backend.

We have created a rake task to sync backend users to the back office (https://github.com/DEFRA/waste-carriers-back-office/pull/182) but this will require an RFC every time it is run. So it's better to hand that power over to the back office super users, so they can easily add a new user and make sure they have access to both apps, without having to wait for the rake task to be run.

* Set up models for backend user data

There are three collections of backend user data we need to deal with - agency_users, admins and roles. This commit creates models and factories for all 3 so we can test the current lib script more closely once we move it into a service.

* Copy lib script into service and test it

This commit creates a new UserMigrationService task containing the contents of lib/users/sync_users.rb. The script is unmodified, except for the way we set database collections on initialising - this moves in some logic which was outsourced to a separate Db module before (also in lib).

We also add tests to make sure that the service runs as expected in its current form.

* Update rake task to use service

`rake users:sync_internal_users` now makes use of the service class instead of the lib class. This means we can delete the file from lib.

* Use nested module/class definitions

Our code style prefers Module Foo Class Bar over Class Foo::Bar, so we need to update our backend models.

* Remove additional files from lib

These are no longer needed now that the previous user-syncing script has been moved.

* Get instance vars for user groups from models

Previously we were getting the collections from the database. This commit switches to getting the lists of objects by calling `Model.all` instead.

* Test the value of @results

Add specs to make sure the correct data is added to @results when a record is created, updated or skipped.

* Improve tests for role determination

Add tests to make sure we correctly assess all the different types of roles for agency_users and admins. This highlighted a few inaccuracies with the models, which are also fixed here.

* Add role_has_changed? method

Just a minor bit of abstraction to improve readability.

* Rename determined_role to backend_role

We pass this variable around a lot, so hopefully this will make what it is a bit clearer.

* Remove instance vars for object lists

Setting these instance variables (eg @agency_users) made sense when we were loading in database collections - however, now that we're just relying on the model and only use each instance variable once, it adds unnecessary complexity.

* Add comments explaining use of Mongo methods

Our usage of `insert_one` and `set` is a bit unorthodox, so added an explanation for why we've done it this way.

* Change how we access model attributes

Switch from `user[:attribute]` to `user.attribute` as this is more standard across the application. Previously we had to use the hash-like syntax as we were loading data straight from the collection, but now that we are using the models, we can change this.

* Shift logic around to reduce complexity and duplication

This is a minor change, but takes a bit of the work away from the `sync_admin` and `sync_agency` methods, which removes a small amount of duplication.

* Add controller for user migrations

Plus views, routes and tests. Only super-users should be able to access these actions.

* Redirect to results page after posting

Posting to the create route for migrations should redirect to the results page afterwards, to avoid refreshing repeating the post request.

* Run user migrations on post request

When the create action is triggered, cause a new UserMigrationService to sync the users.

* Pass results of sync to results page

After running the sync, get the results and pass them to the results page so we can display the output.

* Add content to migration start page

Explain what running the migration will and will not do, and add a button to run the migration.

* Display results of migration

On the migration results page, display a total count of each result (created, skipped, etc) and list users who were created, updated or raised errors.

We don't list skipped users as this would be a very long list, and if nothing has changed, there's nothing to show.

* Add link to user management page

Link from the user management page to the migration tool.