Improve dashboard search performance (#191)
https://eaflood.atlassian.net/browse/WC-531

In production our users have reported that when attempting to apply a filter in the dashboard, it becomes unresponsive and eventually leads to a "service unavailable" page.

Initial investigation found that the problem is each time a user searches the app is pulling back all transient registrations and then iterating through them, checking for the matching criteria.

This change is about improving the performance of the search in the dashboard, such that it no longer times out, and is a scalable solution i.e. we do not want to be back here again in another 6 months.

[PR #302](https://github.com/DEFRA/waste-carriers-renewals/pull/302) in waste-carriers-renewals added [named scopes](https://docs.mongodb.com/mongoid/master/tutorials/mongoid-queries/#named-scopes) to the transient registration.

This change takes advantage of them by updating what was the `TransientRegistrationSearchService` (now `TransientRegistrationSearchService`) to convert options selected in the dashboard UI to named scopes. These are then selected and merged into a single criteria, which is used when we finally query the database.

On top of that as Kaminari supports mongoid, we can properly integrate with it to also limit the number of results we get back to match the current page size and number.

Finally more out of reducing the noise in the log, but also to squeeze every bit of performance out, we have reduced the log level for Mongoid.