Add workflow states to conviction_sign_offs (#196)
* Add workflow states to conviction_sign_offs

https://eaflood.atlassian.net/browse/WC-483

The current implementation of the conviction checking workflow allows a flagged renewal to be checked, then approved or rejected. However, the actual process is more complex, and a renewal which does have matching convictions goes through an internal process before a final decision is made. To better represent this process, we have updated the engine to include workflow states for the conviction_sign_off.

This commit updates the back office convictions process to use those states.

* Update call to approve conviction_sign_off

The old `approve` method has been replaced by a new one in the state machine, but we need to call `approve!` in order ot persist the changes.

* Test that workflow_state is updated by controllers

Make sure that POSTing the conviction approval or rejection forms updates the workflow_state for the conviction_sign_off.

* Call state machine when rejecting conviction_sign_off

Instead of the controller doing the work of rejecting the conviction_sign_off, we can now ask the state machine to deal with that work instead.