Add convictions overview page (#197)
* Add convictions overview page

https://eaflood.atlassian.net/browse/WC-483

Back office users can currently check for convictions which require review by searching and filtering on the dashboard. However, the actual workflow is more complex, so we need a separate page with more features to support this.

* Create new page at /bo/convictions

This will become the dashboard for managing the convictions workflow.

* Link to convictions dashboard from menu

* Display renewals requiring checks on conviction dashboard

List all transient_registrations which have pending conviction_sign_offs on the dashboard, and link to the conviction pages.

* List items on convictions dashboard by last modified

The most recently updated items should appear at the top.

This should fix some intermittently failing tests.

* Refactor dashboard results into shared code

Since we will soon need views for renewals at different stages of the conviction workflow, refactor the bits we'll be reusing into reusable code.

* Add possible matches page to convictions dashboard

This page lists only renewals which have been flagged as potentially having conviction matches and require an initial review for approval or further checks.

* Add checks in progress page to convictions dashboard

This page lists only renewals which have been flagged as requiring further processing after the initial check.

* Add approved checks page to convictions dashboard

This page lists only renewals which have been approved during a convictions check.

* Add rejected checks page to convictions dashboard

This page lists only renewals which have been rejected during a convictions check.

* Add links between dashboard views

Add a mini menu to the dashboard to allow users to view renewals at different stages of the conviction check process.

* Display possible matches on index

After further thinking, I'm not convinced there is a user need for a page listing any renewal which was ever flagged for a conviction check regardless of the current status of that check. So instead we display the possible matches on the index and remove the old possible matches page.

* Display message on conviction dashboard when no results found

Make it clearer that the service has no results to list so it doesn't look like an error has occured or the service has timed out.