Remove conviction status from dashboard search filters (#203)
* Remove conviction status from dashboard search filters

https://eaflood.atlassian.net/browse/WC-483

With the addition of the new convictions dashboard (https://github.com/DEFRA/waste-carriers-back-office/pull/197) we no longer need to include conviction status as a filter on the main search dashboard. This PR removes that functionality.

* Remove filter from dashboard view

Remove the checkbox which allows users to activate the filter.

* Remove conviction filter from TransientRegistrationSearchService

* Remove conviction filter from DashboardsController

As we no longer accept anything from the view to pass to the service.