Refactor AdminFormsController options (#209)
An upcoming change will require us to pass an additional param to the AdminFormsController's `create` method. To stop the number of arguments from getting too high, this PR implements a new `options` hash instead, which should allow us greater flexibility in what options we pass to the AdminFormsController.

I did this for the `new` method as well for consistency's sake. Also moved the authorization logic into its own method to reduce duplication and improve clarity.