Add status in dashboard for stuck renewals (#217)
https://eaflood.atlassian.net/browse/WC-533

WC-533 is linked to one of a number of examples we've had where renewals have become stuck. In all cases it was because the renewal was expired, however when we attempted to complete the renewal we hadn't spotted this leading to a failed completion.

In these cases the renewals have no pending payments or convictions, and have been submitted. Therefore the current dashboards view doesn't display a status.

To better highlight those that are stuck, and to indicate to the user that action needs to be taken to resolve it, this changes updates the view to ensure a status is shown for stuck renewals.

Handily, this will also highlight those that were rejected at conviction check, as we do not have a solution for how to progress them after this implemented as yet.