Rescue CanCan::AccessDenied by redirecting to correct page (#222)
Resolves https://github.com/DEFRA/waste-carriers-back-office/issues/181

The location of the error pages in the engine was modified, but we didn't update the back-office CanCan error handling to deal with this. This PR fixes the error handling to redirect to the permissions error page, instead of a page which doesn't exist (resulting in a 404 error page instead).