Add test coverage for correct handling of 404 errors
Part of: https://eaflood.atlassian.net/browse/RUBY-261

This adds test coverage for correct handling of 404 errors.
The fix on the engine as not been applied but looks like this project is
not affected by the issue. Yet I think is best to push this in for
both consistency and to prevent regression.