Fix persistence of updated session token after logout (#328)
https://eaflood.atlassian.net/browse/RUBY-381

This fix worked with our Postgres apps but didn't seem to persist with Mongoid. I think this is because Mongoid was attempting to re-validate the password when we persisted the session token. Using `set` gets around that by only storing the update to the session token and not bothering with the other attributes.