Add missing companies house configuration (#361)
https://eaflood.atlassian.net/browse/RUBY-659

Our experience building Waste Carriers, Waste Exemptions and Flood Risk Activity Exemptions is that they generally share the same validations.

To reduce duplication and help maintenance we built the [defra-ruby-validators](https://github.com/DEFRA/defra-ruby-validators) gem. One of its validators is to confirm that for LTD/LLP organisations the companies house number provided exists and belongs to an active entity.

We've recently incorporated the gem in Waste Carriers, but it expects that at somepoint a call will be made to configure it, because host apps must provide the companies house API key.

After some investigation we've managed to find that we've not done this, which means the validator is not working, which for LTD/LLP users means their registration will be blocked.