Include registrations in dashboard search (#382)
https://eaflood.atlassian.net/browse/RUBY-704

The back office dashboard search was originally designed to only look at renewals - records in the TransientRegistrations table. We want to look at the Registrations table as well so we can see completed registrations and new applications in progress.

This will require updating our search service, as well as eventually changing the design of the dashboard to be able to handle both types of object and show the correct information and actions for them.

This PR does the work of updating the search so that we get the registrations into the results, although they're not displayed the way that we ultimately want them to be. For a quick fix, we hide the status flags and actions for registrations.