Add 'inactive' tag text (#391)
https://eaflood.atlassian.net/browse/RUBY-696

Our tags didn't account for the existence of the 'inactive' status, which some existing registrations have. This was causing an I18n error, which this PR fixes.

We're going with the same colour as 'revoked' and 'rejected' for now as this status can be applied if a user ceases their own registration.