Display correct date on search results (#387)
https://eaflood.atlassian.net/browse/RUBY-704

All search results should have something in the 'date' column, but what this is and how it is displayed varies based on the result type and its current status.

We also hide the date if inactive, revoked or refused. Initially we planned to show the date this happened. However we don't store that. So for now, leave it blank.