Add registration details page (#445)
* Add registration details page 

This PR contains all the necessary code for showing a details page for registration. 
A presenter has been used instead of a helper in the view. This is because we have already previously decided that we wanted a presenter for the details page as the business logic is very complex. I had to cut refactoring somewhere, hence the presenter will need some further work to move more of the view logic into it. 
There are still some outstanding questions regarding: 
 - Show `in progress` / `active` status somewhere. Will need to speak with Martin. 
 - I don't know which page the conviction checks link should redirect to. Will need to speak with Alan.