Add ceased/revoked info box to view details pages (#460)
https://eaflood.atlassian.net/browse/RUBY-754

If a resource has been ceased or revoked, we want to display some information about this on the 'view details' page.

This should include whether it is considered ceased or revoked, who did it and when, and what the reason was. The reason should only be visible for certain user types.