Update ActionLinksHelper to use cleaner methods (#471)
This came out of trying to fix https://github.com/DEFRA/waste-carriers-back-office/pull/470. I noticed we had some methods here that were still checking the object class and not allowing for it to be a presenter. It didn't end up being part of the solution but I thought we should change it anyway.

We also have nicer methods for checking the metadata status, so I added those too.