Redirects after a renewing is confirmed (#476)
* Redirects after a renewing is confirmed

From: https://eaflood.atlassian.net/browse/RUBY-782

When confirming a payment or approving convictions, a renewing registration might get confirmed and evolve to a registration. In that case, the details page of the renewing registration will not exist any more. Historically, we used to just redirect the user to the main dashboard, but during one of the many refactoring we have killed that functionality, and the test suite didn't highlight the issue due to having different mongoId settings than other environments. Hence, I have also killed those special settings as part of this story.
Now, if a user gets to the renewal details page with a renewal id that no longer exist, we will check if we have a registration matching that id and we will redierct the user there. If there is not even a registration matching the given id, we will redirect to the dashboard as we were historically doing.