Remove all links to old backend (#478)
https://eaflood.atlassian.net/browse/RUBY-786

Since the two system do not share login cookies, when moving from the new back office to the old back office users are asked to login again every time. In the case of starting a new registration, they were blocked to do so half way through the process and redirected to login and then to the admin's dashboard.

Since there is currently no functionality on the old code to deal with the redirect to the correct page after login, we decided to kill all the links for our initial release.