Deal with new frontend regs on possible matches list (#485)
https://eaflood.atlassian.net/browse/RUBY-791

Any conviction_sign_offs that are created in the waste-carriers-frontend don't get a workflow_state when they are created. So we need to call an additional scope to include those as well. This will make sure that new, unapproved conviction_sign_offs are still listed in the 'possible matches' list when appropriate.