Create conviction approval form for registrations (#492)
https://eaflood.atlassian.net/browse/RUBY-792

Until we have moved new registrations into the engine and they become transient_registrations, we need to have a slightly separate process for approving convictions on registrations.

This PR sets up a version of the conviction approval form that works with pending new registrations. It doesn't cover making the approval actually do anything.

We can reuse the ConvictionApprovalForm, but a separate controller, route and view is required.

This is a bit cludgy and duplicates a lot of code, but the intent is that it's very easy to remove once we no longer need it!