Fix broken back link on successful reg transfer (#512)
This path hadn't been updated after the tweaks in https://github.com/DEFRA/waste-carriers-back-office/pull/502. As it was, it was causing the page to error. Updating the path fixes it.

I was wondering why this wasn't caught by the test suite and then realised there were no request tests for the success page. So this also adds some specs!

This led to spotting that there was no permissions check on the success page, so let's fix that too.