Update user list (#513)
https://eaflood.atlassian.net/browse/RUBY-759

Currently, the back office has users and a screen which lists them. However it's just a read-only view. As we move more user functionality in the back office, we need to update the user dashboard to include them.

This PR adds additional columns and links for the user status and actions. These links don't do anything yet, but that will change when we implement future stories.