Add base boxi report generator service (#525)
* Add base boxi report generator service

From: https://eaflood.atlassian.net/browse/RUBY-738

Since with MongoDb all our entities are embedded documents, in order to optimise our algorithm we have to change approach from what we have done with WEX. We only want to fetch all registrations once, and for each of them we will add an entry to all the correct files. This is achieved through an additional service which will run the query and deleage to each parser the job of adding the correct entries to the respective files.
An integration test for the BOXI export functionality has been added as a result of this.
The different file serializers will come in separate PRs.