Change back office user role (#569)
https://eaflood.atlassian.net/browse/RUBY-761

It should be possible for super users to change the roles of other back office users.

This is dependent on the user group - for example, an agency super can only change the role of other agency users. The roles which are available to choose are also limited by the super user's group - so a finance super can only select finance roles.

This also includes an abilities refactor. We can now have just one `:modify_user` ability with checks based on the user role. Worth noting that I couldn't get doubles to work with the CanCanCan tests, so unfortunately the unit test is not as pure as we'd ideally like.

Finally, we make some changes to our password validation. This adopts some fixes we made in WEX, including improved password validation messages, and only running the validation method when the password is submitted through the form.

Devise handles password length validation for us with :validatable, so we don't need to add our own validation here. We also don't need to attempt to validate the format when no password is provided as Devise should handle this. This matches what we did on WEX and the advice from the Devise wiki: https://github.com/plataformatec/devise/wiki/How-To:-Set-up-simple-password-complexity-requirements