Add 'invite new user' feature (#574)
https://eaflood.atlassian.net/browse/RUBY-760

Back office super users should be able to invite new back office users.

Users will be expected to enter the email address for the new user, select a role, and when submitted behind the scenes the app will create a new user and send an email with a link to the email. The link validates the email address and presents a screen that asks the user to create a password.

Super users should only be able to invite users with roles of their user group - for example, agency supers can invite new agency users, but not new finance users.