Fix default email protocol (#590)
https://eaflood.atlassian.net/browse/RUBY-760

During QA of the 'invite user' feature, we spotted that the email confirmation link did not work as expected when opened directly into incognito mode.

This turned out to be because of the default protocol being `http`. This may just be an issue on our test environments, but we should default to `https` anyway.