Add 'developer' user role (#592)
https://eaflood.atlassian.net/browse/RUBY-772

We want to be able to update our conviction data from within the app. We'll need a new type of user group for this, called 'developer'.

Developer users will have the same permissions as agency_users to start out with, with one exception. So we'll just consider them a part of the agency user group.

They also have the `:import_conviction_data` ability. We won't use this for anything yet, but it's unique to developer users and the reason for the role's existence.