Set up ConvictionImportService (#594)
https://eaflood.atlassian.net/browse/RUBY-772

Currently our convictions data is updated using the old waste-carriers-service in Java. We want to decomission that so we need a way for our new codebase to update the data.

This PR sets up a new ConvictionImportService which is designed to take raw data and update our list of convictions in the database.

We also want to avoid deleting the old convictions data and then raising an error, leaving the collection empty. Ideally we would do all of this in a transaction to avoid that problem. But Mongoid 5 doesn't have them! (6 does though.) The best thing we can do is just hold off on making any database changes until we are as sure as possible that the new data is valid.