Fix parsing carriage returns in BOXI export (#634)
https://eaflood.atlassian.net/browse/RUBY-738

PR #625 tackled a number of issues including finding new lines in the text we are trying to export as CSV. However, it only looked for `"\n"` whereas subsequent tests of the export have highlighted it's actually full-blown carriage returns (`"\r\n"`) we have to deal with. This is no doubt a result that most users of our system are on Windows machines.

So this change is a slight tweak to the existing parser. Rather than getting too clever, we replace the existing `gsub` with a function built into Rails for just this purpose. So now we `squish` em!