View registration certificate in back office (#642)
https://eaflood.atlassian.net/browse/RUBY-834

We already email out certificates from the back office with completed renewals.

However, the old backend also has a feature where you can view a certificate on-demand through the back office. The PDF generation is already in place in the engine. We just need to load it from a route.

This is basically copied from what we do to render PDFs in the WEX back office - in that case, confirmation letters and renewal letters.

One difference is we have to require wicked_pdf in the gemfile. This isn't necessary in WEX but without it WCR doesn't render the PDF - instead we get the certificate in HTML no matter what. So may be an issue with the engine config. But this gets us green tests for now.