Fix finance role permissions (#653)
https://eaflood.atlassian.net/browse/RUBY-884

We need to make a number of fixes to the permissions for our finance roles.

Notably:

- The `finance_user` type exists solely to deal with bank transfer payments
- `finance_super` should be able to do everything a `finance-admin` can do, plus the extra user management bits
- `agency_user`s should not be able to do anything financial

This PR should update all our permissions to match the spreadsheet.

Also tried to refactor the tests to be more like the agency ones. However it isn't as clear because the `finance` role is still off doing its own weird thing.