Finance - Charge adjust (#644)
* Finance - Charge adjust 

From: https://eaflood.atlassian.net/browse/RUBY-870

This PR contains the code for positive and negative charge adjust.
Some explanations on the choices I have made:

I have used forms which currently inherits from the Engine base form, even though I only need basic setups and I do nothing with the resource object itself. In the future, if we generate resources forms, we can decide to inherit from a simpler model.
I have decided to not reuse exactly the logic to validate and cast a string to integer used for the other payments form. This is because in the case when a user insert invalid characters that system currently fails with a 500 error while the old system has a custom validation of Invalid format, hence with Andrew we agreed we would validate the format as per old system and keep the field as text.
There is a start_form for the charge adjust, but no state machine. I have decided that would be an overkill, as I don't really need to persist the positive VS negative choice hence I don't need a transient object. However, it is arguable that the controller looks a bit hacky as it does not use the standard Resource form submit implementation. Happy to discuss / change approach to it :)