Fix bug on worlpday missed payment for registrations (#674)
* Fix bug on worlpday missed payment for registrations

From: https://eaflood.atlassian.net/browse/RUBY-826

We used to run a little piece of code that would fix the workflow state of a blocked renewal when adding a worldpay missed payment. That code will not work when run on a registration (nor there is the need to run it there). Hence, this PR adds a check that will make sure we only run that code with transient objects which do have a workflow state to work with.