Redirect to correct page if renewal completes (#670)
https://eaflood.atlassian.net/browse/RUBY-895

We have realised there is an issue when we perform an action that causes a renewal to complete, like taking a payment. We would attempt to redirect back to the finance details for the renewal, even though we just deleted it!

The primary focus of this change is solving that issue. It acknowledges that all controllers that inherit from `ResourceFormsController` could cause a renewal to complete. So we've moved the `CanRenewIfPossible` there. We have also added a new `renew_if_possible_and_redirect()` in the `ResourceFormsController` to fix our issue.

This also includes some additional refactoring, mainly around charge adjustments, to bring it inline with our other payment actions.

We have realised that once a renewal has been completed we have the need to redirect the user to the correct resource page, as the transient resource will be deleted at that point.
This contains the refactoring and fixes necessary to make that happen.