Add ability to complete a registration after payment (#680)
* Add ability to complete a registration after payment

From: https://eaflood.atlassian.net/browse/RUBY-826

We have realised that we don't only have to renew a registration when a payment clears the balance on a renewal, but also complete a registration when the payments clears the balance of a registration and there are no pending conviction checks. This updates the code to handle the completion of a registration when a payment is taken.