Fix validation on amounts (#689)
From: https://eaflood.atlassian.net/browse/RUBY-894

This PR fixes some insconsistency on the way we validate our amounts in the payments and charge adjust forms.
This also fixes a bug for which submitting a payment form without an amount results in a 500 error. This bug was dues
to the fact that we were not validating the form data before creating a payment on the object when submitting the form
.
Another issue related to this has to deal with an existing bug on the re-implementation of the Enumerable module from
rails, for which the `sum` method called on a `[nil]` collection returns 0 rather than throwing an error. Given this,
our test suite was missing to report the error because our test always run on objects with no existing payments.