Fix renewal bug on charge adjust (#702)
* Fix renewal bug on charge adjust

From: https://eaflood.atlassian.net/browse/RUBY-870

We have realised that by running the service for charge adjust in the controller, we are in fact updating the finance details after the renewal completion service had the chance of running.
In order to make sure the service is run before we check for the possibility of completing or renewing a resources, we have taken advantage of Callback. We also renamed the callback to better represent it.