Clarify payment details access (#726)
https://eaflood.atlassian.net/browse/RUBY-908

A recent PR (#717) was an initial attempt to resolve an issue found with renewals and finance actions. If any finance action, such as a charge adjust, is made before the renewal application is complete (`submitted`) it will be wiped at the end of the application process.

It was agreed in the team that rather than try to handle in-complete renewals in these transactions (as we are not aware of a need to do so) we'd instead only allow them to be performed on registrations that have a status of `submitted`.

PR #717 focused on payments, whereas what we really want to hide is the link to the payment details screen altogether, until such time as the renewal application is complete.

This PR is about updating the code to match this expected behaviour.