Remove Cease & revoke link from dashboard (#728)
https://eaflood.atlassian.net/browse/RUBY-931

It seems a while ago now but we have implemented a new Cease & Revoke function in the back office to replace the existing de-register and revoke functions in the old backend.

This has yet to be shipped, so before we do we'd like to amend a feature of the new implementation. When it was added to the back office it followed the convention in the old code of adding a link to perform the task directly in the dashboard results.

However, we are trying to simplify things as much as possible and that includes reducing the number of things you see in the dashboard results, especially listing actions that are also available in the registrations details screen.

This change covers removing it from the back-office dashboard page.