Remove Transfer link from dashboard page (#729)
https://eaflood.atlassian.net/browse/RUBY-932

When we first shipped the back-office it included a new ‘Transfer’ function to replace the broken one in the backend. However, you accessed it from a link in the backend because the back-office had no registration details view.

Fast forward to the recent tech debt work, and the ‘Registration details’ screen was added, along with seeing registrations in the back-office dashboard results. With them in place, we also removed the ‘Transfer’ link from the backend and made it visible in the back-office.

When the dashboard link was added to the back office it followed the convention in the old code of adding a link to perform the task directly in the dashboard results.

However, we are trying to simplify things as much as possible and that includes reducing the number of things you see in the dashboard results, especially listing actions that are also available in the registrations details screen.