Fix conviction approval access permissions (#746)
https://eaflood.atlassian.net/browse/RUBY-945

It has been spotted in recent testing that the permissions around convictions do not match with expectations. In the old service agency & agency with refund users could approve conviction checks. After working with NCCC it has been agreed that this should be restricted to just agency with refund, and agency super users.

However, currently, it seems all users can see the details, and all agency users (so not finance) can approve conviction checks.

This change is about ensuring the system behaves as expected.

** Notes

Having touched the conviction request tests this change also incorporates some refactoring of them. To improve test speed we're now happy to use multiple `expect` statements for `request` specs. So this includes a number of changes to merge some of the tests in the request specs.