Add view details page for new registrations (#878)
https://eaflood.atlassian.net/browse/RUBY-1054

We want NCCC staff to be able to access details of a new in-progress registration, as they can with renewals.

This PR sets up a version of the 'view details' page for in-progress registrations. This page should be linked to in the relevant search results.

We also don't want to include any finance info on this page.