Add rake task to remove old transient_registrations (#886)
https://eaflood.atlassian.net/browse/RUBY-10

We've agreed that we shouldn't store transient_registrations for more than 30 days. So we need a way of clearing out the old ones.

This rake task will find and remove transient_registrations that are older than 30 days. It should also remove all related objects (eg transient_addresses).

We don't delete submitted renewals. In case a renewal is held up by a pending conviction check or payment, we want to make sure we don't delete it.