Fix defra-ruby-features ActiveRecord error (#927)
https://eaflood.atlassian.net/browse/RUBY-1118

We think there is an issue when using our new [defra-ruby-features](https://github.com/DEFRA/defra-ruby-features) engine in a WCR 'production' environment. It seems to be related to an `ApplicationRecord` class in the engine that isn't actually used.

So we have [updated the features engine](https://github.com/DEFRA/defra-ruby-features/pull/1) to remove the file. This updates the back-office to use that latest version.