Remove link to certificate from renewal complete (#945)
https://eaflood.atlassian.net/browse/RUBY-1154

The link in the page takes you to the frontend where if you are not signed in, you then have to in order to access the link. We don't provide this link on the new 'new registration' done page. It is also our intention to remove accounts in the near future so access to this will be removed anyway.

So this change is just about removing the link in the page to make it consistent with the new registration complete page, and with our future intentions. The change was actually done in the engine as [PR #887](https://github.com/DEFRA/waste-carriers-engine/pull/887). We are just updating the version of the engine to bring the change into the back-office.