Generate renewal token when one is requested (#962)
https://eaflood.atlassian.net/browse/RUBY-1166

If a registration is eligible for renewal NCCC would like us to display the same renewal link that gets included in the renewal reminder emails and letters in registration details.

Currently, though we only generate the token used in the link when first generating the reminder email. A registration is eligible for renewal before this is done. And when we first release there will be hundreds that are eligible which will never go through the reminder process because it is too late.

The simplest solution we have for this is to generate it if required. We made that change in the engine in [PR #898](https://github.com/DEFRA/waste-carriers-engine/pull/898). This change updates the back-office to use the new functionality.