Update flash message logic (#973)
https://eaflood.atlassian.net/browse/RUBY-1168
https://eaflood.atlassian.net/browse/RUBY-1169

When resending the renewal reminder email from the registration details screen it has been flagged its doesn't really work from a UI perspective if the contact email is missing or the AD default.

So we have an update to the design which requires us to show a flash message styled as an error. Initially, it looked like we didn't support this. But turns out that is due to the differences between WEX and WCR and how flash messages have been implemented.

So to prevent anyone else getting confused, this brings WCR into line with WEX.

Note. The use of a `shared/_flash_messages.html.erb` partial is different from WEX. But we will look to

- move the partial into the [waste-carriers-engine](https://github.com/DEFRA/waste-carriers-engine)
- add the same partial to [waste-exemptions-engine](https://github.com/DEFRA/waste-exemptions-engine) and refactor the views to use it