Only display renew/new reg links to those who can use them (#975)
https://eaflood.atlassian.net/browse/RUBY-1176

Previously we displayed this "Renew" button to all back office users, including ones who didn't have permission.

This PR changes it to use a different method which checks permissions as well as the state of the registration.

While in here, I also noticed a TODO which could now be resolved. So if it is now impossible to renew, but the user can make a new reg instead, we'll display a link to that.