Show err msg when resend reminder to empty contact (#972)
https://eaflood.atlassian.net/browse/RUBY-1168

Currently if you click the 'Resend renewal reminder' link on a registration with an empty contact email it will just say 'Renewal sent to'.

With this message you could be mistaken for thinking it sent successfully. Or if you spot it, it wouldn;t be immediately obvious what the issue is.

We now have a design to display an error message if the link is clicked and the contact email is blank. This change implements it.

* Add exceptions module with first error

Opted to store our exceptions in a central place as per https://stackoverflow.com/a/5369773/6117745.

Advice on Rails 6 now is also to store the stuff you would have previously placed in `lib/` in `app/lib` as per https://github.com/rails/rails/issues/37835

Because we are stil using the classic autoload it seems we needed to manually add `app/lib` to the autoload_path. But in future when the app is ready to switch to the new Zeitwerk loader no other changes should be needed.

* Update renewal_reminder_mailer to throw exception

We had 2 options really

- check for the presence of contact_email before calling one of the methods in renewal_reminder_mailer
- throw an exception in  renewal_reminder_mailer if contact_email is missing

The second option means we only need to do the check in one place, but each caller can handle the exception as they see fit.

* Update resend renewal email controller

Update the controller to handle the new exception and display an appropriate error message. This includes updating how unexpected exceptions are reported.

* Update tests to cover new behaviour

Update of the tests to cover the different flash messages depending on the scenario, plus a spot of refactoring.